sqlite: store if coin is from self
This adds a new `is_from_self` column to the coins table that
indicates if the coin is an output of a transaction for which
all inputs are ours. In the case of unconfirmed coins, being
from self requires that all unconfirmed ancestors, if any, are
also from self.

The DB migration sets the value of this column for all existing
coins and the poller keeps the column up to date.