gui: improve home reload logic
If the user has selected an event when the wallet finishes
syncing, don't reload the page so that the selected event is
not cleared.

When reloading the page, clear any selected event even if the
wallet is syncing so that the user is not stuck with the event
until the wallet has finished syncing.