poller: pass `LianaDescriptor` instead of single descs
The poller already assumes we are dealing with a receive and
change descriptor, rather than an arbitrary number of single
descriptors, so this does not really limit its generality.

As such, it seems more straightforward to pass a single
`LianaDescriptor` rather than an array with fixed length 2,
especially as we will want to pass the receive and change
descriptors to other methods (specifically those relating to
setting receive and change indices in the DB).